model/CRNN/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
backports=1.0=py_2
backports.weakref=1.0rc1=py36_0
blas=1.0=mkl
bleach=1.5.0=pypi_0
ca-certificates=2020.1.1=0
certifi=2020.4.5.1=py36_0
cudatoolkit=8.0=3
cudnn=6.0.21=cuda8.0_0
cycler=0.10.0=pypi_0
decorator=4.4.2=pypi_0
easydict=1.6=pypi_0
html5lib=0.9999999=pypi_0
imageio=2.8.0=pypi_0
intel-openmp=2020.0=166
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc=7.2.0=h69d50b8_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libprotobuf=3.11.4=hd408876_0
libstdcxx-ng=9.1.0=hdf63c60_0
markdown=3.1.1=py36_0
matplotlib=2.0.2=pypi_0
mkl=2019.5=281
mkl-service=2.3.0=py36he904b0f_0
mkl_fft=1.0.15=py36ha843d7b_0
mkl_random=1.1.0=py36hd6b4f25_0
ncurses=6.2=he6710b0_0
networkx=2.4=pypi_0
numpy=1.18.2=pypi_0
numpy-base=1.18.1=py36hde5b4d6_1
opencv-python=3.2.0.7=pypi_0
openssl=1.1.1=h7b6447c_0
pillow=7.0.0=pypi_0
pip=20.0.2=py36_1
protobuf=3.11.4=py36he6710b0_0
pyparsing=2.4.6=pypi_0
python=3.6.10=h0371630_0
python-dateutil=2.8.1=pypi_0
pytz=2019.3=pypi_0
pywavelets=1.1.1=pypi_0
readline=7.0=h7b6447c_5
scikit-image=0.16.2=pypi_0
scipy=1.4.1=pypi_0
setuptools=46.0.0=py36_0
six=1.14.0=py36_0
sqlite=3.31.1=h7b6447c_0
tensorflow-gpu=1.3.0=pypi_0
tensorflow-gpu-base=1.3.0=py36cuda8.0cudnn6.0_1
tensorflow-tensorboard=0.1.8=pypi_0
tk=8.6.8=hbc83047_0
tqdm=4.43.0=pypi_0
werkzeug=1.0.0=py_0
wheel=0.34.2=py36_0
xz=5.2.4=h14c3975_4
zlib=1.2.11=h7b6447c_3
